# openapi.yaml
openapi: 3.0.1
info:
  title: Delilah Resto
  description: API Restaurant.
  contact:
    email: a-d-r-u@hotmail.com
  version: 1.0.0
...
servers:
  - url: 'https://localhost:3000'
  - url: 'http://localhost:3000'
...
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get list of all available products
      description: Returns a list of products with corresponding information
      responses:
        '200':
          description: successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Products'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Products
      summary: Adds a new product (Admin)
      description: Adds a product to database
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        '200':
          description: New product added
        '400':
          description: Error validating input data
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Default'
...
  /orders:
    get:
      tags:
        - Orders
      summary: Get list of orders
      description: >-
        Returns a list of orders. You can only access to your own orders or if you are an admin you can access to all the ordesr.
      responses:
        '200':
          description: successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orders'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        default:
          $ref: '#/components/responses/Default'
    post:
      tags:
        - Orders
      summary: Adds a product
      description: Adds a product to database
      requestBody:
        description: Order to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewOrder'
        required: true
      responses:
        '200':
          description: Order modified succesfully
        '400':
          description: Error validating input data
        '401':
          $ref: '#/components/responses/Unauthorized'
        default:
          $ref: '#/components/responses/Default'
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        product_id:
          type: integer
          example: 2
        name:
          type: string
          example: Caldo de Costilla
        price:
          type: number
          example: 8000
        is_disabled:
          type: boolean
          example: 0
    Products:
      type: array
      items:
        $ref: '#/components/schemas/Product'
    OrderProduct:
      type: object
      properties:
        order_prod_id:
          type: integer
          example: 16
        order_id:
          type: integer
          example: 15
        product_id:
          type: integer
          example: 4
        product_amount:
          type: integer
          example: 3
        name:
          type: string
          example: Hamburguesa Vegetariana
        price:
          type: number
          example: 25000
        is_disabled:
          type: boolean
          example: 0
    OrderProducts:
      type: array
      items:
        $ref: '#/components/schemas/OrderProduct'
    Order:
      type: object
      properties:
        order_id:
          type: integer
          example: 1
        status:
          type: string
          example: new
        date:
          type: string
          format: date-time
          example: '2020-10-11T10:11:12Z'
        description:
          type: string
          example: 1x Caldo de Costilla
        payment_method:
          type: string
          example: cash
        total:
          type: integer
          example: 8000
        user_id:
          type: integer
          example: 1
        username:
          type: string
          example: Adriana
        full_name:
          type: string
          example: Adriana Ariza
        email:
          type: string
          example: adr@123.com
        phone:
          type: integer
          example: 7562341
        delivery_address:
          type: string
          example: Calle123
        products:
          $ref: '#/components/schemas/OrderProducts'
    Orders:
      type: array
      items:
        $ref: '#/components/schemas/Order'
    NewOrder:
      type: object
      properties:
        payment_method:
          type: string
          example: cash
        data:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: integer
                example: 1
              amount:
                type: integer
                example: 3
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: 'An error has ocurred: <error>'
      required:
        - error
    Send:
      type: string
      example: Error message
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: The specified resource was not found
    Default:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/Send'
          examples:
            Token Invalidation:
              summary: Unauthorized - Invalid Token
              value: Unauthorized - Invalid Token
            Disabled Account:
              summary: Unauthorized - Account is Disabled
              value: Unauthorized - User account is disabled
...